+++ src/main.lua
@@ -1,5 +1,6 @@
----@meta _
+-- TODO, clean this shit UP, this file SUCKS so much.
 
+---@meta _
 ---@diagnostic disable-next-line: undefined-global
 local mods = rom.mods
 
@@ -19,6 +20,7 @@ modutil = mods["SGG_Modding-ModUtil"]
 chalk = mods["SGG_Modding-Chalk"]
 reload = mods["SGG_Modding-ReLoad"]
 sjson = mods["SGG_Modding-SJSON"]
+---@module 'zannc-GodsAPI-auto'
 gods = mods["zannc-GodsAPI"].auto()
 
 ---@module 'Droppable_Gods-zannc-config'
@@ -61,9 +63,11 @@ local function on_ready()
 	StackingandCostForGods(game.TraitData, "NPC_Artemis_Field_01")
 	StackingandCostForGods(game.TraitData, "NPC_Athena_01")
 	StackingandCostForGods(game.TraitData, "NPC_Dionysus_01")
+	StackingandCostForGods(game.TraitData, "NPC_Hades_Field_01")
 
 	game.EnemyData.NPC_Dionysus_01.AlwaysShowDefaultUseText = true
 	game.EnemyData.NPC_Athena_01.AlwaysShowDefaultUseText = true
+	game.EnemyData.NPC_Hades_Field_01.AlwaysShowDefaultUseText = true
 
 	-- Super Magical Code by Jowday, which doesn't give weird codex unlock popup
 	modutil.mod.Path.Wrap("AttemptOpenUpgradeChoiceBoonInfo", function(base, screen, button)
@@ -88,6 +92,13 @@ local function on_ready()
 				},
 			}
 			base(newScreen, button)
+		elseif screen.Source.Name == "zannc-Droppable_Gods-HadesUpgrade" then
+			local newScreen = {
+				Source = {
+					Name = "NPC_Hades_Field_01",
+				},
+			}
+			base(newScreen, button)
 		else
 			base(screen, button)
 		end
@@ -97,6 +108,7 @@ local function on_ready()
 	game.EnemyData.NPC_Artemis_Field_01.WeaponUpgrades = {}
 	game.EnemyData.NPC_Athena_01.WeaponUpgrades = {}
 	game.EnemyData.NPC_Dionysus_01.WeaponUpgrades = {}
+	game.EnemyData.NPC_Hades_Field_01.WeaponUpgrades = {}
 
 	if config.Artemis.enabled then
 		table.insert(godList, "NPC_Artemis_01")
@@ -110,12 +122,52 @@ local function on_ready()
 		table.insert(godList, "NPC_Dionysus_01")
 		import("gods/dionysusLoot.lua")
 	end
+	if config.Hades.enabled then
+		table.insert(godList, "NPC_Dionysus_01")
+		import("gods/hadesLoot.lua")
+		-- if not game.CodexData.OlympianGods.Entries.NPC_Hades_Field_01 then
+		-- 	return
+		-- end
+		-- game.CodexData.OlympianGods.Entries.NPC_Hades_Field_01.NoRequirements = false
+	end
 
-	for k, npcVal in pairs(godList) do
+	local pMenu = mods["PonyWarrior-PonyMenu"]
+	for _, npcVal in pairs(godList) do
 		if not game.CodexData.OlympianGods.Entries[npcVal] then
 			return
 		end
 		game.CodexData.OlympianGods.Entries[npcVal].NoRequirements = false
+		if pMenu then
+			for _, v in ipairs(pMenu.CommandData) do
+				if v.Name == "NPC_Artemis_Field_01" then
+					v.NoSpawn = false
+				end
+				if v.Name == "NPC_Athena_01" then
+					v.NoSpawn = false
+				end
+				if v.Name == "NPC_Dionysus_01" then
+					v.NoSpawn = false
+				end
+				if v.Name == "NPC_Hades_Field_01" then
+					v.NoSpawn = false
+				end
+			end
+		end
+	end
+
+	if pMenu then
+		modutil.mod.Path.Wrap("CreateLoot", function(base, args)
+			if args.Name == "NPC_Artemis_Field_01" then
+				args.Name = "zannc-Droppable_Gods-ArtemisUpgrade"
+			elseif args.Name == "NPC_Athena_01" then
+				args.Name = "zannc-Droppable_Gods-AthenaUpgrade"
+			elseif args.Name == "NPC_Dionysus_01" then
+				args.Name = "zannc-Droppable_Gods-DionysusUpgrade"
+			elseif args.Name == "NPC_Hades_Field_01" then
+				args.Name = "zannc-Droppable_Gods-HadesUpgrade"
+			end
+			return base(args)
+		end)
 	end
 end
 
@@ -145,6 +197,7 @@ modutil.once_loaded.game(function()
 		["zannc-Droppable_Gods-ArtemisUpgrade"] = "NPC_Artemis_Field_01",
 		["zannc-Droppable_Gods-AthenaUpgrade"] = "NPC_Athena_01",
 		["zannc-Droppable_Gods-DionysusUpgrade"] = "NPC_Dionysus_01",
+		["zannc-Droppable_Gods-HadesUpgrade"] = "NPC_Hades_Field_01",
 	}
 
 	for k, godName in pairs(npcGods) do